feat(kafka): extend EmailsScored and EmailsVerdict for pipeline spine
Add component score pointers, partial-analysis fields, ComponentDetails
(raw JSON per analyser), and verdict explainability fields. Update
contract tests and fix notification service logging for int risk_score.

+++ services/svc-09-notification/cmd/notification/main.go
@@ -40,7 +40,7 @@ func handle(ctx context.Context, msg kafkaconsumer.Message, _ svckit.Deps) error
 	zerolog.Ctx(ctx).Info().
 		Int64("email_id", v.Meta.EmailID).
 		Str("verdict", v.VerdictLabel).
-		Float64("risk_score", v.RiskScore).
+		Int("risk_score", v.RiskScore).
 		Msg("would-notify")
 	return nil
 }
